Move the `smoke` test plan to a nested directory
